flake8, move cut_plotter to views
